Update setup.py
ERROR: After October 2020 you may experience errors when installing or updating packages. This is because pip will change the way that it resolves dependency conflicts. We recommend you use --use-feature=2020-resolver to test your packages with the new resolver before it becomes the default.

django-py3-ldap 1.0.0 requires ldap3==2.5, but you'll have ldap3 2.8.1 which is incompatible.
django-py3-ldap 1.0.0 requires pyasn1==0.4.5, but you'll have pyasn1 0.4.8 which is incompatible.